frontend/marketplace/backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
azure-functions
google-generativeai
azure-cosmos>=4.3.0
azure-storage-blob>=12.14.0
azure-identity>=1.5.0
requests>=2.25.1
requests-toolbelt>=0.9.1
azure-data-tables>=12.0.0
azure-cognitiveservices-speech
python-barcode
reportlab